Reorganized the util package
- WARNING: This means every plugin depending on PcCore must be recompiled! No code changes needed, just an adjustment of the imports.

- Added source encoding (UTF-8) to pom
