+++ src/main/java/com/perceivedev/perceivecore/command/CommandTree.java
@@ -17,8 +17,8 @@ import org.bukkit.plugin.Plugin;
 
 import com.perceivedev.perceivecore.PerceiveCore;
 import com.perceivedev.perceivecore.language.MessageProvider;
-import com.perceivedev.perceivecore.util.DisableListener;
 import com.perceivedev.perceivecore.util.DisableManager;
+import com.perceivedev.perceivecore.util.types.DisableListener;
 
 /** A Command tree */
 public class CommandTree {
