+++ src/main/java/com/perceivedev/perceivecore/config/util/DataManager.java
@@ -1,6 +1,6 @@
 package com.perceivedev.perceivecore.config.util;
 
-import static com.perceivedev.perceivecore.util.TextUtils.normalizePathName;
+import static com.perceivedev.perceivecore.util.text.TextUtils.normalizePathName;
 
 import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
