+++ src/main/java/com/perceivedev/perceivecore/particle/effect/shapes/Line3D.java
@@ -5,7 +5,7 @@ import org.bukkit.Particle;
 import org.bukkit.World;
 import org.bukkit.util.Vector;
 
-import com.perceivedev.perceivecore.particle.math.SphericalCoordinates;
+import com.perceivedev.perceivecore.util.math.SphericalCoordinates;
 
 /**
  * A Line, which points from one Vector to another.
