+++ src/main/java/com/perceivedev/perceivecore/util/text/Unicode.java
@@ -1,5 +1,5 @@
 
-package com.perceivedev.perceivecore.util;
+package com.perceivedev.perceivecore.util.text;
 
 public enum Unicode {
